feat : multiRead - 다건 조회 구현 #89

+++ src/main/java/com/midas/shootpointer/domain/post/repository/PostQueryRepository.java
@@ -37,5 +37,8 @@ public interface PostQueryRepository extends JpaRepository<PostEntity,Long> {
                    "LIMIT :size",
             nativeQuery = true
     )
-    List<PostEntity> getPopularPostListBySliceAndNoOffset(@Param(value = "lastPostId")Long lastPostId,@Param(value = "size")int size,@Param(value = "likeCnt")int likeCnt);
+    List<PostEntity> getPopularPostListBySliceAndNoOffset(@Param(value = "lastPostId")Long lastPostId,
+                                                          @Param(value = "size")int size,
+                                                          @Param(value = "likeCnt")Long likeCnt
+    );
 }
